mysite/requirements.txt
mysqlclient==1.3.12
pymysql==0.9.2
Pillow==5.2.0
djangorestframework==3.8.2
django-filter==2.0.0
django-bootstrap-form==2.1.1
django-widget-tweaks==1.4.3
django-guardian
